[CHORE] #93 : Indicator 컴포넌트 불리
